+++ app/assets/javascripts/analyses/index.js
@@ -1,25 +1,26 @@
-$(function() {
-  const showDialog = function(id) {
-    const dialog = new bs.Modal(document.getElementById(id));
-    dialog.show();
-  };
+import * as common from '../application';
 
+$(function() {
   $('#nav-link-analysis').addClass('active');
 
+  document.querySelectorAll('[data-bs-toggle="tooltip"]').forEach(function (element) {
+    new bs.Tooltip(element);
+  });
+
   $('#new_analysis').on('ajax:success', function(event) {
-    showDialog('dialog-execute');
+    common.showDialog('dialog-execute');
   });
 
   $('#new_analysis').on('ajax:error', function(event) {
-    showDialog('dialog-execute-error');
+    common.showDialog('dialog-execute-error');
   });
 
   $('#table-analysis').on('ajax:success', '.rebuild', function(event) {
-    showDialog('dialog-execute');
+    common.showDialog('dialog-execute');
   });
 
   $('#table-analysis').on('ajax:error', '.rebuild', function(event) {
-    showDialog('dialog-execute-error');
+    common.showDialog('dialog-execute-error');
   });
 
   $('#analysis_data_source').on('change', function(event) {
@@ -37,21 +38,7 @@ $(function() {
   });
 
   $('tbody').on('click', '.btn-param', function(event) {
-    const analysisId = $(this).parents('tr').attr('id');
-    $.ajax({
-      type: 'GET',
-      url: `/alterf/api/analyses/${analysisId}/parameter`,
-    }).done(function(parameter) {
-      $('#parameter-max_depth').text(parameter.max_depth || '指定なし');
-      $('#parameter-max_features').text(parameter.max_features);
-      $('#parameter-max_leaf_nodes').text(parameter.max_leaf_nodes || '指定なし');
-      $('#parameter-min_samples_leaf').text(parameter.min_samples_leaf);
-      $('#parameter-min_samples_split').text(parameter.min_samples_split);
-      $('#parameter-num_tree').text(parameter.num_tree);
-      showDialog('dialog-parameter');
-    }).fail(function(xhr, status, error) {
-      showDialog('dialog-parameter-error');
-    });
+    common.showParameterDialog($(this).parents('tr').attr('id'));
   });
 
   $('#tbody-analysis').on('click', '.download', function(event) {
@@ -68,7 +55,7 @@ $(function() {
         $('<a>', {href: blobUrl, download: filename})[0].click();
         (URL || webkitURL).revokeObjectURL(blobUrl);
       } else {
-        showDialog('dialog-download-error');
+        common.showDialog('dialog-download-error');
       }
     };
     xhr.send();
@@ -86,7 +73,7 @@ $(function() {
     }).done(function(data) {
       location.href = `/alterf/analyses?${$.param(queries)}`;
     }).fail(function(xhr, status, error) {
-      showDialog('dialog-execute-error');
+      common.showDialog('dialog-execute-error');
     });
   });
 });