+++ app/assets/javascripts/evaluations/manage.js
@@ -1,7 +1,8 @@
+import * as common from '../application';
+
 $(function() {
   const formCollapse = document.getElementById('form-evaluation');
-
-  $('#nav-link-evaluation').addClass('active');
+  const collapse = new bs.Collapse(formCollapse);
 
   formCollapse.addEventListener('show.bs.collapse', function(event) {
     $('button#collapse-form > span')
@@ -13,6 +14,28 @@ $(function() {
     $('button#collapse-form > span')
       .removeClass('bi-dash-circle')
       .addClass('bi-plus-circle');
+  });
+
+  $('#collapse-form').on('click', function() {
+    collapse.toggle();
+  });
+
+  $('#new_evaluation').on('ajax:success', function(event) {
+    common.showDialog('dialog-execute');
+  });
+
+  $('#new_evaluation').on('ajax:error', function(event) {
+    common.showDialog('dialog-execute-error');
+  });
+
+  $('#table-evaluation').on('ajax:error', function(event) {
+    common.showDialog('dialog-download-error');
+  });
+
+  $('#nav-link-evaluation').addClass('active');
+
+  document.querySelectorAll('[data-bs-toggle="tooltip"]').forEach(function (element) {
+    new bs.Tooltip(element);
   });
 
   $('#evaluation_data_source').on('change', function(event) {
@@ -34,11 +57,20 @@ $(function() {
     if (state === 'waiting' || state === 'error') {
       return;
     }
+
+    if ($(this).attr('class') === 'model' && event.target.tagName === 'BUTTON') {
+      return;
+    }
+
     if ($(this).attr('class') === 'download') {
       return;
     }
 
     open(`/alterf/evaluations/${row.attr('id')}`, '_blank');
+  });
+
+  $('#table-evaluation').on('click', '.model button', function(event) {
+    common.showParameterDialog($(this).attr('id'));
   });
 
   $('#table-evaluation').on('ajax:success', function(event) {
@@ -49,5 +81,19 @@ $(function() {
     const filename = /filename="(.*)"/.exec(xhr.getResponseHeader('Content-Disposition'))[1];
     $('<a>', {href: blobUrl, download: filename})[0].click();
     (URL || webkitURL).revokeObjectURL(blobUrl);
+  });
+
+  $('#table-evaluation').on('mouseover', '.download button, .model button', function(event) {
+    const rowId = $(this).parents('tr').attr('id');
+    const row = document.getElementById(rowId);
+    const tooltip = bs.Tooltip.getInstance(row);
+    tooltip.hide();
+    tooltip.disable();
+  }).on('mouseleave', '.download button, .model button', function(event) {
+    const rowId = $(this).parents('tr').attr('id');
+    const row = document.getElementById(rowId);
+    const tooltip = bs.Tooltip.getInstance(row);
+    tooltip.enable();
+    tooltip.show();
   });
 });