+++ app/channels/evaluation/datum_channel.rb
@@ -1,5 +1,0 @@
-class Evaluation::DatumChannel < ApplicationCable::Channel
-  def subscribed
-    stream_from 'evaluation_datum'
-  end
-end