+++ app/models/denebola/race.rb
@@ -1,4 +1,3 @@
-module Denebola
-  class Race < Denebola::Base
-  end
+class Denebola::Race < Denebola::Base
+  has_many :entries, dependent: :nullify
 end