+++ db/migrate/20210714142647_rename_tree_id_column_to_analysis_result_decision_trees.rb
@@ -1,0 +1,5 @@
+class RenameTreeIdColumnToAnalysisResultDecisionTrees < ActiveRecord::Migration[6.1]
+  def change
+    rename_column :analysis_result_decision_trees, :tree_id, :decision_tree_id
+  end
+end