+++ spec/controllers/api/analyses/show_spec.rb
@@ -25,7 +25,7 @@ describe Api::AnalysesController, type: :controller do
         end,
         decision_trees: analysis.result.decision_trees.map do |decision_tree|
           {
-            tree_id: decision_tree.tree_id,
+            decision_tree_id: decision_tree.decision_tree_id,
             nodes: decision_tree.nodes.map do |node|
               node.slice(
                 :node_id,
@@ -57,9 +57,6 @@ describe Api::AnalysesController, type: :controller do
   describe '異常系' do
     before { @analysis_id = 'not_exist' }
     include_context 'リクエスト送信'
-
-    it 'ステータスコードが正しいこと' do
-      is_asserted_by { @response_status == 404 }
-    end
+    it_behaves_like 'ステータスコードが正しいこと', 404
   end
 end