+++ spec/factories/analysis/result/decision_tree.rb
@@ -1,6 +1,6 @@
 FactoryBot.define do
   factory :decision_tree, class: 'Analysis::Result::DecisionTree' do
-    tree_id { 0 }
+    decision_tree_id { '000000' }
     nodes { [build(:node)] }
   end
 end